[brain tests] Burst enum tests from the main file
